Change Layout test to be Router test
